fixed small error in landscape script
in line 101, small error in the very last bit where it should be Min(size -1, i + step) instead of Min(size-1, i + 1).

Also, remove all doubles as I dont think its very needed, so instead of Math, I switch it to Mathf where it is for floats. All I did was add an 'f' in every 'Math' word.

I changed rng.nextDouble() to Random.value because it is essentially the same thing except this is for floats

I also added walls in the scene